move error to errors module
